FB-15a: extract attemptApiRequest and recursivelyMakeDiracRequests into TaskRequestLoop.ts
